Revert "Revert "Revert "kljlkjd"""
This reverts commit 2fcaaeeb3122467deadf99e7183735f5353f9f90.

+++ Project_SoCMainFrame/os_yuntai.h
@@ -10,10 +10,10 @@ class Service_Yuntai
 {
 public:
 	Service_Yuntai();
-	static long long int yuntai_init();
-	static long long int yuntai_quick(long long int a,long long int b);
-	static long long int yuntai_slow(long long int a,long long int b);
-	static long long int yuntai_stop(long long int a,long long int b);
+	static int yuntai_init();
+	static int yuntai_quick(int a,int b);
+	static int yuntai_slow(int a,int b);
+	static int yuntai_stop(int a,int b);
 
 private:
 };
